Deleted a scary amount of unused ffi/native cpp code
